Take a big brush and clean this project up
* Added "deps" directory - symlink openssl there according to README
* Removed compilation warnings (except for blake_ref - it will be removed
  in upcoming commit anyway)
* Unified project settings through all configurations
* Moved to more common /bin and /obj build folder scheme
* Moved projects to their separate directories in src

+++ src/BlakeHasher/blake_ref.cpp
@@ -436,7 +436,7 @@ static HashReturn Final32(hashState * state, BitSequence * hashval)
     u32 low, high;
     low = state->t32[0] + state->datalen;
     high = state->t32[1];
-    if (low < state->datalen)
+    if (low < static_cast<u32>(state->datalen))
         high++;
     U32TO8_BE(msglen + 0, high);
     U32TO8_BE(msglen + 4, low);
