+++ src/R41N30W/adriansReduction.cpp
@@ -4,12 +4,12 @@
 static const char mCharset[] = "abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ0123456789._";
 static const unsigned int mCharsetLength = 64;
 
-void AdrianReduction(const int salt, const int resultLength, const ucharVectorPtr& hashValue, ucharVectorPtr& plainValue)
+void AdrianReduction(const int salt, const size_t resultLength, const ucharVectorPtr& hashValue, ucharVectorPtr& plainValue)
 {
     // clear() leaves capacity unchanged - no need to reserve
     plainValue->clear();
 
-    for (int i = 0; i < resultLength; i++)
+    for (size_t i = 0; i < resultLength; i++)
     {
         unsigned int index = (*hashValue)[i] % mCharsetLength;
         plainValue->push_back(mCharset[index]);
